Fixed issue where deleting worker database which doesnt exist raises

+++ packages/syft/src/syft/core/node/worker.py
@@ -2,6 +2,7 @@
 from __future__ import annotations
 
 # stdlib
+import contextlib
 from functools import partial
 import hashlib
 import os
@@ -184,7 +185,9 @@ class Worker(NewNode):
         if reset:
             store_config = SQLiteStoreClientConfig()
             store_config.filename = f"{uid}.sqlite"
-            os.unlink(store_config.file_path)
+            with contextlib.suppress(FileNotFoundError):
+                if os.path.exists(store_config.file_path):
+                    os.unlink(store_config.file_path)
         return Worker(name=name, id=uid, signing_key=key, processes=processes)
 
     def is_root(self, credentials: SyftVerifyKey) -> bool:
